feat(bottom-sheet): changed currency type of value

+++ app/src/main/java/com/paixao/dev/gastu/presentation/model/HomeModel.kt
@@ -1,14 +1,15 @@
 package com.paixao.dev.gastu.presentation.model
 
 import com.paixao.dev.gastu.domain.entities.HomeEntity
+import java.math.BigDecimal
 
 data class HomeModel constructor(
-    val currency: Float,
-    val spendCurrency: Float,
-    val earningCurrency: Float,
+    val currency: BigDecimal,
+    val spendCurrency: BigDecimal,
+    val earningCurrency: BigDecimal,
     val deals: List<DealModel>?
 )
 
 internal fun HomeEntity.toModel(): HomeModel {
-    return HomeModel(currency, spendCurrency, earningCurrency, deals?.map { it.toModel() })
+    return HomeModel(currency.toBigDecimal(), spendCurrency.toBigDecimal(), earningCurrency.toBigDecimal(), deals?.map { it.toModel() })
 }
